requirements.txt
argparse>=1.1
requests>=2.21.0
gensim>=3.6.0
nltk>=3.2.5
numpy>=1.18.2
pandas>=1.0.3
bcolz>=1.2.1
pyspellchecker>=0.5.4
spacy>=2.2.4
tqdm>=4.38.0
scikit-learn>=0.22.2
torch>=1.4.0